fix comments & depend on sdk 6.6.0 aar library.
