+++ Vungle/build.gradle
@@ -7,6 +7,6 @@ dependencies {
         implementation project(':vungle-android-sdk:publisher-sdk-android')
     } else {
         //implementation "com.vungle:publisher-sdk-android:${sdkVersion}"
-        implementation 'com.github.vungle:vungle-android-sdk:6.5.0-RC4'
+        implementation 'com.github.Vungle:vungle-android-sdk:6.6.0-QA2'
     }
 }
