+++ Vungle/src/main/java/com/mopub/mobileads/VungleBanner.java
@@ -12,7 +12,6 @@
 
  import com.mopub.common.logging.MoPubLog;
  import com.mopub.common.util.Views;
- import com.mopub.mobileads.VungleRewardedVideo.VungleMediationSettings;
  import com.vungle.warren.AdConfig;
  import com.vungle.warren.AdConfig.AdSize;
  import com.vungle.warren.VungleNativeAd;
@@ -128,8 +127,8 @@
 
         sVungleRouter.addRouterListener(mPlacementId, mVungleRouterListener);
 
-        VungleMediationSettings.adConfigWithLocalExtras(adConfig, localExtras);
-        if (VungleMediationSettings.isStartMutedNotConfigured(localExtras)) {
+        VungleMediationConfiguration.adConfigWithLocalExtras(adConfig, localExtras);
+        if (VungleMediationConfiguration.isStartMutedNotConfigured(localExtras)) {
             adConfig.setMuted(true); // start muted by default
         }
