+++ Vungle/src/main/java/com/mopub/mobileads/VungleInterstitial.java
@@ -7,7 +7,6 @@ import androidx.annotation.Keep;
 import androidx.annotation.NonNull;
 
 import com.mopub.common.logging.MoPubLog;
-import com.mopub.mobileads.VungleRewardedVideo.VungleMediationSettings;
 import com.vungle.warren.AdConfig;
 
 import java.util.Map;
@@ -118,7 +117,7 @@ public class VungleInterstitial extends CustomEventInterstitial {
 
         if (localExtras != null) {
             mAdConfig = new AdConfig();
-            VungleMediationSettings.adConfigWithLocalExtras(mAdConfig, localExtras);
+            VungleMediationConfiguration.adConfigWithLocalExtras(mAdConfig, localExtras);
         }
 
         sVungleRouter.loadAdForPlacement(mPlacementId, mVungleRouterListener);
